requirements.txt
# requirements.txt
fastapi
strawberry-graphql
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
python-dotenv

# Para pruebas
pytest
httpx
